# openapi.yaml
openapi: 3.0.3
info:
  title: BerryFi Studio API
  description: |-
    This document provides the official API specification for the BerryFi Studio Pixel Streaming Platform.
    It is intended for backend developers to implement the required endpoints for the frontend application.
  version: 1.0.0
  contact:
    name: Shubham Bokade
    email: bloggersbokade17@gmail.com
servers:
  - url: https://api.berryfi.studio/v1
    description: Production server for BerryFi Studio API
...
components:
...
  schemas:
...
    User:
      type: object
      description: Represents a user account in the system.
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier for the user.
          example: "user_mithesh"
          readOnly: true
        name:
          type: string
          description: The full name of the user.
          example: "Mithesh Bhat"
        email:
          type: string
          format: email
          description: The user's unique email address, used for login.
          example: "mithesh@ravgroup.org"
        role:
          type: string
          enum: [super_admin, org_owner, org_admin, org_auditor, org_reporter, org_billing, org_member, workspace_admin, workspace_auditor, workspace_reporter, workspace_billing, workspace_member]
          description: The role of the user, which determines their permissions.
          example: "org_owner"
        accountType:
          type: string
          enum: [organization, workspace]
          description: The type of account the user belongs to.
          example: "organization"
        organizationId:
          type: string
          description: The organization ID for the user.
          example: "ravgroup"
        workspaceId:
          type: string
          description: The workspace ID for the user (if applicable).
          example: "apexmarketing"
        status:
          type: string
          enum: [active, invited, disabled]
          description: The current status of the user account.
          example: "active"
        lastLogin:
          type: string
          format: date-time
          description: The timestamp of the user's last successful login.
          readOnly: true
          example: "2024-01-15T10:30:00Z"

    # ====== Request/Response Schemas ======
    LoginRequest:
      type: object
      description: Credentials required for a user to log in.
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "admin@berryfi.com"
        password:
          type: string
          format: password
          description: User's password. Must be at least 8 characters.
          minLength: 8
          example: "SuperSecurePassword@123"
    AuthResponse:
      type: object
      description: The response payload after a successful authentication, containing tokens and user info.
      properties:
        user:
          $ref: "#/components/schemas/User"
        accessToken:
          type: string
          description: A short-lived JWT (JSON Web Token) for authenticating subsequent API requests.
          example: "mock_jwt_token_user_mithesh_1705312200000"
        refreshToken:
          type: string
          description: A long-lived token used to obtain a new access token without requiring the user to log in again.
          example: "mock_refresh_token_user_mithesh_1705312200000"
    RefreshTokenRequest:
      type: object
      description: The refresh token required to obtain a new access token.
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: The long-lived refresh token obtained during login.
          example: "mock_refresh_token_user_mithesh_1705312200000"
    RefreshTokenResponse:
      type: object
      description: The response payload after a successful token refresh.
      properties:
        accessToken:
          type: string
          description: A new, short-lived JWT for API authentication.
          example: "mock_jwt_token_refreshed_1705312200000"

    # ====== Standard Error Schema ======
    ApiError:
      type: object
      description: A standard error response format for client (4xx) and server (5xx) errors.
      required:
        - statusCode
        - message
      properties:
        statusCode:
          type: integer
          format: int32
          description: The HTTP status code.
          example: 400
        message:
          type: string
          description: A human-readable description of the error.
          example: "Invalid email format"
        error:
          type: string
          description: A short, machine-readable error string.
          example: "Bad Request"

  # ============================
  # ====== Security Schemes ======
  # ============================
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Authorization header using the Bearer scheme. The token is a short-lived **access token** obtained from the `/auth/login` or `/auth/refresh` endpoints."

# ============================
# ====== Global Security ======
# ============================
security:
  - BearerAuth: [] # Apply Bearer Authentication globally to all endpoints by default.

# ============================
# ====== API Paths & Endpoints ======
# ============================
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: User Login
      description: Authenticates a user with their email and password. Returns user details, a short-lived `accessToken`, and a long-lived `refreshToken`.
      operationId: loginUser
      security: [] # This endpoint does not require an access token.
      requestBody:
        description: User credentials for login.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad Request. The request body is malformed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthorized. The provided email or password is incorrect.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
                example:
                  statusCode: 401
                  message: "Invalid email or password."
                  error: "Unauthorized"
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh Access Token
      description: Obtains a new `accessToken` by providing a valid `refreshToken`. This is used to maintain a user's session without requiring them to log in again.
      operationId: refreshAccessToken
      security: [] # This endpoint does not require an access token.
      requestBody:
        description: The long-lived refresh token obtained from a successful login.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Access token refreshed successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        "401":
          description: Unauthorized. The provided refresh token is invalid, expired, or has been revoked. The user must log in again.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get Current User
      description: Retrieves the profile of the currently authenticated user based on the provided `accessToken`.
      operationId: getCurrentUser
      responses:
        "200":
          description: User profile retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized. The `accessToken` is missing, malformed, or expired. The client should attempt to use the `refreshToken` to get a new one.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Forbidden. The user's account is disabled or lacks permission for this action, even with a valid token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/logout:
    post:
      tags: [Authentication]
      summary: User Logout
      description: Logs out the current user. The backend should invalidate the `refreshToken` to prevent it from being used to generate new access tokens.
      operationId: logoutUser
      responses:
        "204":
          description: Logout successful. No content is returned.
        "401":
          description: Unauthorized. The request was made without a valid `accessToken`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /projects/{projectId}:
    get:
      tags: [Projects]
      summary: Get Project Details
      description: Get detailed information about a specific project.
      operationId: getProject
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
          example: "river_castle_project"
      responses:
        '200':
          description: Project details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      description:
                        type: string
                      status:
                        type: string
                      productionUrl:
                        type: string
                      config:
                        type: object
                      branding:
                        type: object
                      links:
                        type: object
  /projects/{projectId}/deploy:
    post:
      tags: [Projects]
      summary: Deploy Project
      description: Deploy a project to production.
      operationId: deployProject
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project deployed successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
  /projects/{projectId}/stop:
    post:
      tags: [Projects]
      summary: Stop Project
      description: Stop a running project.
      operationId: stopProject
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project stopped successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
...
  /projects/{projectId}/status:
    get:
      tags: [Projects]
      summary: Get Project Status
      description: Get the current status of a project.
      operationId: getProjectStatus
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        example: "running"
                      lastDeployed:
                        type: string
                        format: date-time
                      currentCCU:
                        type: integer
                        example: 15
                      uptime:
                        type: number
                        example: 99.5
                      errors:
                        type: array
                        items:
                          type: string
...