# src/main/java/com/berryfi/portal/dto/auth/RefreshTokenRequest.java
package com.berryfi.portal.dto.auth;

import jakarta.validation.constraints.NotBlank;

/**
 * DTO for refresh token request.
 */
public class RefreshTokenRequest {

    @NotBlank(message = "Refresh token is required")
    private String refreshToken;

    public RefreshTokenRequest() {}

    public RefreshTokenRequest(String refreshToken) {
        this.refreshToken = refreshToken;
    }

    public String getRefreshToken() {
        return refreshToken;
    }

    public void setRefreshToken(String refreshToken) {
        this.refreshToken = refreshToken;
    }
}