# src/main/java/com/berryfi/portal/dto/auth/RefreshTokenResponse.java
package com.berryfi.portal.dto.auth;

/**
 * DTO for refresh token response.
 */
public class RefreshTokenResponse {

    private String accessToken;

    public RefreshTokenResponse() {}

    public RefreshTokenResponse(String accessToken) {
        this.accessToken = accessToken;
    }

    public String getAccessToken() {
        return accessToken;
    }

    public void setAccessToken(String accessToken) {
        this.accessToken = accessToken;
    }
}